Move `configure` to authentication file
